Restructuration du code
