+++ app/__init__.py
@@ -1,2 +1,20 @@
+'''
 from flask import Flask
 from .views import app, socket_io
+'''
+
+from flask import Flask
+from flask_socketio import SocketIO
+
+flaskapp = Flask(__name__, instance_relative_config=True)
+socket_io = SocketIO()
+print("Initialisation des apps..;")
+
+def create_app(debug=False):
+    flaskapp.config.from_object('config')
+    flaskapp.debug = debug
+
+    import app.main
+
+    socket_io.init_app(flaskapp)
+    return flaskapp
