+++ app/main/__init__.py
@@ -1,0 +1,2 @@
+from . import routes, events
+print("Yo")
