+++ run.py
@@ -1,6 +1,7 @@
-from app import socket_io
-from app import app
+from app import create_app, socket_io
+
+app = create_app(True)
 
 if __name__ == "__main__":
-    socket_io.run(app, host='0.0.0.0')
+    socket_io.run(app)
